Don't print out that null is an object
Previously if you passed "null" for a typed field Anchor would return the very
confusing error message

    `undefined` should be a text (instead of "null", which is a object)

Instead now we return the better error message:

    `foo` should be a number (instead of null)

Note this requires you to pass the key `validation` as part of the `context`.
Waterline doesn't do this, yet, but it will soon.

Adds a simple test for both the null case and the regular case.