Don't return an array of errors
If more than one thing breaks, too bad; just return the first validation error,
instead of all of them. This will simplify the programming model and the error
handling model in Waterline/Sails significantly.

Furthermore, properly initialize and pass back Error objects. The previous
error constructor incorrectly initialized the Error constructor with a
dictionary instead of an error message, which would lead to Error: [object
Object] being created instead of an error message.

We should be throwing errors instead of returning an Error or null, but that
will have to come in a later commit since this is big enough already.