Return a slightly better message for invalid rules
Really, each rule in lib/match/rules.js should throw an error on invalid input,
so you can say things like "The maximum allowed value is 5 (input: 6)" or
whatever, instead of having one rule apply to everything.

But this should be a slight improvement - specifically, we now include the
column name in the error message, if we have it. An error message that looked
like this:

    "min" validation rule failed for input: -1

Now looks like this:

    Invalid day. Input failed min validation: -1

Still not great, but better, and includes the column name.