Special case 'required' rule
It comes up often enough that it's worth adding a better error message for this
case - in this instance, "No email was provided. Please provide a email" is
better than "Invalid name. Input failed required validation: ''", when
"required validation" can read as one phrase instead of "required" as the
validation type.

Adds better formatting in lib/match/rules.js and removes a lot of the tabbing
in that file.